+++ apps/decodex/src/recovery/tests/stale_active/reentry.rs
@@ -1,482 +1,4 @@
-use std::fs;
-
-use tempfile::TempDir;
-
-use crate::{
-	recovery::{
-		GHOST_LANE_TERMINAL_STATUS, RUN_CONTROL_CHANNEL_STATUS_FAILED,
-		RecoveryRuntimeMutationPolicy, STALE_ACTIVE_CLASSIFICATION,
-		tests::{
-			self, GhostLaneTestTracker,
-			stale_active::{self},
-		},
-	},
-	tracker,
-};
-
-#[test]
-fn stale_active_release_allows_reentry_after_local_cleanup_audit() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let context = tests::sample_recovery_context(
-		&temp_dir,
-		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
-	);
-	let active_label = tracker::automation_active_label(context.config.service_id());
-	let queue_label = tracker::automation_queue_label(context.config.service_id());
-	let worktree_path = context.config.worktree_root().join("PUB-1626");
-	let mut issue =
-		tests::sample_issue_with_labels("In Progress", &[active_label.clone(), queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&context.state_store, &issue, &worktree_path);
-
-	context
-		.state_store
-		.update_run_status("run-1626", GHOST_LANE_TERMINAL_STATUS)
-		.expect("run should terminalize");
-	context
-		.state_store
-		.retire_run_control_channel_for_attempt("run-1626", 1, RUN_CONTROL_CHANNEL_STATUS_FAILED)
-		.expect("control channel should retire");
-
-	fs::remove_dir_all(&worktree_path).expect("worktree should be removed");
-
-	context
-		.state_store
-		.clear_worktree_mapping(&issue.id)
-		.expect("issue-id worktree mapping should clear");
-
-	stale_active::append_stale_active_release_audit(&context.state_store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
-	let mut diagnostics = super::diagnose_stale_active_issues(
-		context.config.service_id(),
-		&context.workflow,
-		context.config.worktree_root(),
-		&context.state_store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-	assert!(diagnostic.evidence.contains(&String::from("stale_active_local_cleanup_complete")));
-
-	super::apply_stale_active_release_with_tracker(
-		&tracker,
-		&context.config,
-		&context.workflow,
-		&context.state_store,
-		&diagnostic,
-	)
-	.expect("reentry release should remove active label");
-
-	assert_eq!(
-		tracker.label_removals.borrow().as_slice(),
-		&[(issue.id.clone(), vec![format!("label-{}", active_label.replace(':', "-"))])]
-	);
-	assert_eq!(
-		tracker.state_updates.borrow().as_slice(),
-		&[(issue.id.clone(), String::from("state-todo"))]
-	);
-}
-
-#[test]
-fn stale_active_release_reentry_blocks_active_status_after_local_cleanup_audit() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let context = tests::sample_recovery_context(
-		&temp_dir,
-		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
-	);
-	let active_label = tracker::automation_active_label(context.config.service_id());
-	let queue_label = tracker::automation_queue_label(context.config.service_id());
-	let worktree_path = context.config.worktree_root().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&context.state_store, &issue, &worktree_path);
-
-	context
-		.state_store
-		.update_run_status("run-1626", "running")
-		.expect("run should carry active status");
-	context
-		.state_store
-		.retire_run_control_channel_for_attempt("run-1626", 1, RUN_CONTROL_CHANNEL_STATUS_FAILED)
-		.expect("control channel should retire");
-
-	fs::remove_dir_all(&worktree_path).expect("worktree should be removed");
-
-	context
-		.state_store
-		.clear_worktree_mapping(&issue.id)
-		.expect("issue-id worktree mapping should clear");
-
-	stale_active::append_stale_active_release_audit(&context.state_store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let mut diagnostics = super::diagnose_stale_active_issues(
-		context.config.service_id(),
-		&context.workflow,
-		context.config.worktree_root(),
-		&context.state_store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, "stale_active_recovery_blocked");
-	assert!(!diagnostic.recoverable());
-	assert!(diagnostic.blockers.contains(&String::from("protocol_event_evidence_present")));
-	assert!(diagnostic.blockers.contains(&String::from("child_agent_activity_present")));
-	assert!(diagnostic.blockers.contains(&String::from("protocol_activity_present")));
-}
-
-#[test]
-fn stale_active_release_reentry_terminal_guards_terminal_looking_audited_run() {
-	for status in ["failed", "interrupted"] {
-		let temp_dir = TempDir::new().expect("tempdir should create");
-		let context = tests::sample_recovery_context(
-			&temp_dir,
-			RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
-		);
-		let active_label = tracker::automation_active_label(context.config.service_id());
-		let queue_label = tracker::automation_queue_label(context.config.service_id());
-		let worktree_path = context.config.worktree_root().join("PUB-1626");
-		let mut issue =
-			tests::sample_issue_with_labels("In Progress", &[active_label.clone(), queue_label]);
-
-		issue.identifier = String::from("PUB-1626");
-
-		stale_active::seed_dead_orphan_runtime_telemetry(
-			&context.state_store,
-			&issue,
-			&worktree_path,
-		);
-
-		context
-			.state_store
-			.update_run_status("run-1626", status)
-			.expect("run should carry terminal-looking app-server status");
-		context
-			.state_store
-			.retire_run_control_channel_for_attempt(
-				"run-1626",
-				1,
-				RUN_CONTROL_CHANNEL_STATUS_FAILED,
-			)
-			.expect("control channel should retire");
-
-		fs::remove_dir_all(&worktree_path).expect("worktree should be removed");
-
-		context
-			.state_store
-			.clear_worktree_mapping(&issue.id)
-			.expect("issue-id worktree mapping should clear");
-
-		stale_active::append_stale_active_release_audit(&context.state_store, &issue.id);
-
-		let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
-		let mut diagnostics = super::diagnose_stale_active_issues(
-			context.config.service_id(),
-			&context.workflow,
-			context.config.worktree_root(),
-			&context.state_store,
-			&tracker,
-			Some("PUB-1626"),
-			RecoveryRuntimeMutationPolicy::ReadOnly,
-		)
-		.expect("stale active diagnosis should run");
-		let diagnostic = diagnostics.pop().expect("diagnostic should exist");
-
-		assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-		assert!(diagnostic.recoverable(), "{status} blockers: {:?}", diagnostic.blockers);
-		assert_eq!(diagnostic.latest_attempt_status.as_deref(), Some(status));
-		assert!(diagnostic.evidence.contains(&String::from("stale_active_local_cleanup_complete")));
-
-		super::apply_stale_active_release_with_tracker(
-			&tracker,
-			&context.config,
-			&context.workflow,
-			&context.state_store,
-			&diagnostic,
-		)
-		.expect("reentry release should terminal-guard and remove active label");
-
-		let run = context
-			.state_store
-			.run_attempt("run-1626")
-			.expect("run attempt should read")
-			.expect("run should exist");
-
-		assert_eq!(run.status(), GHOST_LANE_TERMINAL_STATUS);
-		assert_eq!(
-			tracker.label_removals.borrow().as_slice(),
-			&[(issue.id.clone(), vec![format!("label-{}", active_label.replace(':', "-"))])]
-		);
-		assert_eq!(
-			tracker.state_updates.borrow().as_slice(),
-			&[(issue.id.clone(), String::from("state-todo"))]
-		);
-	}
-}
-
-#[test]
-fn stale_active_release_allows_reentry_after_local_cleanup_without_control_channel() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let context = tests::sample_recovery_context(
-		&temp_dir,
-		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
-	);
-	let active_label = tracker::automation_active_label(context.config.service_id());
-	let queue_label = tracker::automation_queue_label(context.config.service_id());
-	let worktree_path = context.config.worktree_root().join("PUB-1626");
-	let mut issue =
-		tests::sample_issue_with_labels("In Progress", &[active_label.clone(), queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	tests::init_clean_git_repo_with_remote_default(&worktree_path, "x/pubfi-pub-1626");
-	stale_active::seed_dead_orphan_runtime_telemetry_without_control_channel(
-		&context.state_store,
-		&issue,
-		&worktree_path,
-	);
-
-	context
-		.state_store
-		.update_run_status("run-1626", GHOST_LANE_TERMINAL_STATUS)
-		.expect("run should terminalize");
-
-	fs::remove_dir_all(&worktree_path).expect("worktree should be removed");
-
-	context
-		.state_store
-		.clear_worktree_mapping(&issue.id)
-		.expect("issue-id worktree mapping should clear");
-
-	stale_active::append_stale_active_release_audit(&context.state_store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
-	let mut diagnostics = super::diagnose_stale_active_issues(
-		context.config.service_id(),
-		&context.workflow,
-		context.config.worktree_root(),
-		&context.state_store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-	assert!(diagnostic.evidence.contains(&String::from("control_channel_missing")));
-	assert!(diagnostic.evidence.contains(&String::from("stale_active_local_cleanup_complete")));
-
-	super::apply_stale_active_release_with_tracker(
-		&tracker,
-		&context.config,
-		&context.workflow,
-		&context.state_store,
-		&diagnostic,
-	)
-	.expect("reentry release should remove active label");
-
-	assert_eq!(
-		tracker.label_removals.borrow().as_slice(),
-		&[(issue.id.clone(), vec![format!("label-{}", active_label.replace(':', "-"))])]
-	);
-}
-
-#[test]
-fn stale_active_release_reentry_without_control_channel_blocks_private_progress() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let context = tests::sample_recovery_context(
-		&temp_dir,
-		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
-	);
-	let active_label = tracker::automation_active_label(context.config.service_id());
-	let queue_label = tracker::automation_queue_label(context.config.service_id());
-	let worktree_path = context.config.worktree_root().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	tests::init_clean_git_repo_with_remote_default(&worktree_path, "x/pubfi-pub-1626");
-	stale_active::seed_dead_orphan_runtime_telemetry_without_control_channel(
-		&context.state_store,
-		&issue,
-		&worktree_path,
-	);
-
-	context
-		.state_store
-		.update_run_status("run-1626", GHOST_LANE_TERMINAL_STATUS)
-		.expect("run should terminalize");
-
-	fs::remove_dir_all(&worktree_path).expect("worktree should be removed");
-
-	context
-		.state_store
-		.clear_worktree_mapping(&issue.id)
-		.expect("issue-id worktree mapping should clear");
-
-	stale_active::append_stale_active_release_audit(&context.state_store, &issue.id);
-	stale_active::append_harness_outcome_with_pr_progress(&context.state_store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let mut diagnostics = super::diagnose_stale_active_issues(
-		context.config.service_id(),
-		&context.workflow,
-		context.config.worktree_root(),
-		&context.state_store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.evidence.contains(&String::from("control_channel_missing")));
-	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_release_reentry_restores_startable_state_after_active_label_release() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let context = tests::sample_recovery_context(
-		&temp_dir,
-		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
-	);
-	let queue_label = tracker::automation_queue_label(context.config.service_id());
-	let worktree_path = context.config.worktree_root().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&context.state_store, &issue, &worktree_path);
-
-	context
-		.state_store
-		.update_run_status("run-1626", GHOST_LANE_TERMINAL_STATUS)
-		.expect("run should terminalize");
-	context
-		.state_store
-		.retire_run_control_channel_for_attempt("run-1626", 1, RUN_CONTROL_CHANNEL_STATUS_FAILED)
-		.expect("control channel should retire");
-
-	fs::remove_dir_all(&worktree_path).expect("worktree should be removed");
-
-	context
-		.state_store
-		.clear_worktree_mapping(&issue.id)
-		.expect("issue-id worktree mapping should clear");
-
-	stale_active::append_stale_active_release_audit(&context.state_store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
-	let mut diagnostics = super::diagnose_stale_active_issues(
-		context.config.service_id(),
-		&context.workflow,
-		context.config.worktree_root(),
-		&context.state_store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
-
-	assert_eq!(
-		diagnostic.classification,
-		super::super::super::STALE_ACTIVE_STATE_RESTORE_CLASSIFICATION
-	);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-	assert!(
-		diagnostic.evidence.contains(&String::from("stale_active_startable_state_restore_pending"))
-	);
-
-	super::apply_stale_active_release_with_tracker(
-		&tracker,
-		&context.config,
-		&context.workflow,
-		&context.state_store,
-		&diagnostic,
-	)
-	.expect("state-restore reentry should complete");
-
-	assert!(tracker.label_removals.borrow().is_empty());
-	assert_eq!(
-		tracker.state_updates.borrow().as_slice(),
-		&[(issue.id.clone(), String::from("state-todo"))]
-	);
-}
-
-#[test]
-fn stale_active_release_reentry_rejects_release_audit_from_other_run() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let context = tests::sample_recovery_context(
-		&temp_dir,
-		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
-	);
-	let active_label = tracker::automation_active_label(context.config.service_id());
-	let queue_label = tracker::automation_queue_label(context.config.service_id());
-	let worktree_path = context.config.worktree_root().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&context.state_store, &issue, &worktree_path);
-
-	context
-		.state_store
-		.update_run_status("run-1626", GHOST_LANE_TERMINAL_STATUS)
-		.expect("run should terminalize");
-	context
-		.state_store
-		.retire_run_control_channel_for_attempt("run-1626", 1, RUN_CONTROL_CHANNEL_STATUS_FAILED)
-		.expect("control channel should retire");
-
-	fs::remove_dir_all(&worktree_path).expect("worktree should be removed");
-
-	context
-		.state_store
-		.clear_worktree_mapping(&issue.id)
-		.expect("issue-id worktree mapping should clear");
-
-	stale_active::append_stale_active_release_audit_for_run(
-		&context.state_store,
-		&issue.id,
-		"run-older",
-		1,
-	);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let mut diagnostics = super::diagnose_stale_active_issues(
-		context.config.service_id(),
-		&context.workflow,
-		context.config.worktree_root(),
-		&context.state_store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(!diagnostic.recoverable());
-	assert!(!diagnostic.evidence.contains(&String::from("stale_active_local_cleanup_complete")));
-	assert!(diagnostic.blockers.contains(&String::from("protocol_activity_present")));
-}
+mod audit_identity;
+mod audited_control_channel;
+mod missing_control_channel;
+mod state_restore;
